+++ ochre/Equipment/PV.py
@@ -104,13 +104,17 @@ class PV(ScheduledLoad):
             if envelope_model is None:
                 raise OCHREException('Must specify PV tilt and azimuth, or provide an envelope_model with a roof.')
             roofs = [bd.ext_surface for bd in envelope_model.boundaries if 'Roof' in bd.name]
-            if not roofs:
-                raise OCHREException('No roofs in envelope model. Must specify PV tilt and azimuth')
-            # Use roof closest to south with preference to west (0-45 degrees)
-            roof_data = pd.DataFrame([[bd.tilt, az] for bd in roofs for az in bd.azimuths], columns=['Tilt', 'Az'])
-            best_idx = (roof_data['Az'] - 185).abs().idxmax()
-            self.tilt = roof_data.loc[best_idx, 'Tilt']
-            self.azimuth = roof_data.loc[best_idx, 'Az']
+            if roofs:
+                # Use roof closest to south with preference to west (0-45 degrees)
+                roof_data = pd.DataFrame([[bd.tilt, az] for bd in roofs for az in bd.azimuths], columns=['Tilt', 'Az'])
+                best_idx = (roof_data['Az'] - 185).abs().idxmax()
+                self.tilt = roof_data.loc[best_idx, 'Tilt']
+                self.azimuth = roof_data.loc[best_idx, 'Az']
+            else:
+                # TODO: convert to self.warn. Need to initialize Simulator first.
+                print('No roofs in envelope model. Defaulting PV tilt to latitude and azimuth to south.')
+                self.tilt = kwargs["location"]["latitude"]
+                self.azimuth = 0
 
         # Inverter constraints
         self.inverter_capacity = inverter_capacity or self.capacity  # in kVA, AC